refactor: rename _file.py to _multimodal.py
The module handles multimodal input detection, content building, and
file type routing — not just file path helpers. _multimodal.py better
reflects its purpose.
