+++ src/openaivec/_multimodal.py
